refactor: centralize converters
