--- OfficeIMO.Converters/OfficeIMO.Converters.csproj
@@ -1,46 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-    <PropertyGroup>
-        <Description>Shared converter interfaces and options for OfficeIMO projects.</Description>
-        <AssemblyName>OfficeIMO.Converters</AssemblyName>
-        <AssemblyTitle>OfficeIMO.Converters</AssemblyTitle>
-
-        <VersionPrefix>1.0.6</VersionPrefix>
-        <TargetFrameworks Condition=" '$([MSBuild]::IsOsPlatform(`Windows`))' ">
-            netstandard2.0;net472;net8.0;net9.0
-        </TargetFrameworks>
-        <TargetFrameworks Condition=" '$([MSBuild]::IsOsPlatform(`OSX`))' Or '$([MSBuild]::IsOsPlatform(`Linux`))' ">
-            net8.0;net9.0
-        </TargetFrameworks>
-        <GeneratePackageOnBuild>False</GeneratePackageOnBuild>
-        <Company>Evotec</Company>
-        <Authors>Przemyslaw Klys</Authors>
-
-        <PackageId>OfficeIMO.Converters</PackageId>
-        <PackageTags>converter;openxml;office;net472;net8.0;net9.0</PackageTags>
-        <PackageProjectUrl>https://github.com/EvotecIT/OfficeIMO</PackageProjectUrl>
-        <PackageLicenseExpression>MIT</PackageLicenseExpression>
-        <DelaySign>False</DelaySign>
-        <IsPublishable>True</IsPublishable>
-        <Copyright>(c) 2011 - 2024 Przemyslaw Klys @ Evotec. All rights reserved.</Copyright>
-        <PackageIcon>OfficeIMO.png</PackageIcon>
-        <RepositoryUrl>https://github.com/EvotecIT/OfficeIMO</RepositoryUrl>
-        <DebugType>portable</DebugType>
-        <ProduceReferenceAssembly>False</ProduceReferenceAssembly>
-        <GenerateDocumentationFile>True</GenerateDocumentationFile>
-        <LangVersion>Latest</LangVersion>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <None Include="..\Assets\OfficeIMO.png">
-            <Pack>True</Pack>
-            <PackagePath>\</PackagePath>
-        </None>
-    </ItemGroup>
-
-    <ItemGroup>
-        <Using Include="System" />
-        <Using Include="System.IO" />
-    </ItemGroup>
-
-</Project>
