+++ OfficeIMO.Examples/Html/Html.ConverterInterface.cs
@@ -1,8 +1,8 @@
 using System;
 using System.IO;
 using System.Text;
-using OfficeIMO.Converters;
 using OfficeIMO.Html;
+using OfficeIMO.Word;
 
 namespace OfficeIMO.Examples.Html {
     internal static partial class Html {
