+++ OfficeIMO.Examples/Html/Html.GenericFont.cs
@@ -1,7 +1,7 @@
 using System;
 using System.IO;
-using OfficeIMO.Converters;
 using OfficeIMO.Html;
+using OfficeIMO.Word;
 
 namespace OfficeIMO.Examples.Html {
     internal static partial class Html {
