+++ OfficeIMO.Examples/Html/Html.Tables.cs
@@ -1,5 +1,5 @@
-using OfficeIMO.Converters;
 using OfficeIMO.Html;
+using OfficeIMO.Word;
 using System;
 using System.IO;
