+++ OfficeIMO.Examples/Markdown/Markdown.ConverterInterface.cs
@@ -1,8 +1,8 @@
 using System;
 using System.IO;
 using System.Text;
-using OfficeIMO.Converters;
 using OfficeIMO.Markdown;
+using OfficeIMO.Word;
 
 namespace OfficeIMO.Examples.Markdown {
     internal static partial class Markdown {
