+++ OfficeIMO.Html/OfficeIMO.Html.csproj
@@ -43,7 +43,6 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\OfficeIMO.Converters\OfficeIMO.Converters.csproj" />
         <ProjectReference Include="..\OfficeIMO.Word\OfficeIMO.Word.csproj" />
     </ItemGroup>
