+++ OfficeIMO.Markdown/Options/MarkdownToWordOptions.cs
@@ -1,5 +1,5 @@
 using System;
-using OfficeIMO.Converters;
+using OfficeIMO.Word;
 
 namespace OfficeIMO.Markdown {
     /// <summary>
