+++ OfficeIMO.Markdown/Options/WordToMarkdownOptions.cs
@@ -1,4 +1,4 @@
-using OfficeIMO.Converters;
+using OfficeIMO.Word;
 
 namespace OfficeIMO.Markdown {
     /// <summary>
