+++ OfficeIMO.Pdf/OfficeIMO.Pdf.csproj
@@ -11,7 +11,6 @@
     <LangVersion>Latest</LangVersion>
   </PropertyGroup>
   <ItemGroup>
-    <ProjectReference Include="..\OfficeIMO.Converters\OfficeIMO.Converters.csproj" />
     <ProjectReference Include="..\OfficeIMO.Word\OfficeIMO.Word.csproj" />
   </ItemGroup>
   <ItemGroup>
