+++ OfficeIMO.Pdf/PdfPageOrientation.cs
@@ -1,0 +1,9 @@
+namespace OfficeIMO.Pdf {
+    /// <summary>
+    /// Specifies page orientation for PDF export.
+    /// </summary>
+    public enum PdfPageOrientation {
+        Portrait,
+        Landscape
+    }
+}
