+++ OfficeIMO.Tests/ConverterRegistryTests.cs
@@ -1,7 +1,7 @@
 using System.IO;
 using System.Text;
-using OfficeIMO.Converters;
 using OfficeIMO.Markdown;
+using OfficeIMO.Word;
 using System.Threading.Tasks;
 using Xunit;
