+++ OfficeIMO.Word/Converters/IWordConverter.cs
@@ -1,7 +1,7 @@
 using System.IO;
 using System.Threading.Tasks;
 
-namespace OfficeIMO.Converters {
+namespace OfficeIMO.Word {
     public interface IWordConverter {
         void Convert(Stream input, Stream output, IConversionOptions options);
         Task ConvertAsync(Stream input, Stream output, IConversionOptions options);
