+++ OfficeImo.sln
@@ -136,8 +136,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OfficeIMO.Html", "OfficeIMO
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OfficeIMO.Markdown", "OfficeIMO.Markdown\OfficeIMO.Markdown.csproj", "{9C2CC069-324D-4B80-88D2-1E3F49FB2C19}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OfficeIMO.Converters", "OfficeIMO.Converters\OfficeIMO.Converters.csproj", "{C12EF28A-AA7C-4965-862A-77A97840C7BD}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -244,18 +242,6 @@ Global
 		{9C2CC069-324D-4B80-88D2-1E3F49FB2C19}.Release|x64.Build.0 = Release|Any CPU
 		{9C2CC069-324D-4B80-88D2-1E3F49FB2C19}.Release|x86.ActiveCfg = Release|Any CPU
 		{9C2CC069-324D-4B80-88D2-1E3F49FB2C19}.Release|x86.Build.0 = Release|Any CPU
-		{C12EF28A-AA7C-4965-862A-77A97840C7BD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C12EF28A-AA7C-4965-862A-77A97840C7BD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C12EF28A-AA7C-4965-862A-77A97840C7BD}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{C12EF28A-AA7C-4965-862A-77A97840C7BD}.Debug|x64.Build.0 = Debug|Any CPU
-		{C12EF28A-AA7C-4965-862A-77A97840C7BD}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{C12EF28A-AA7C-4965-862A-77A97840C7BD}.Debug|x86.Build.0 = Debug|Any CPU
-		{C12EF28A-AA7C-4965-862A-77A97840C7BD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C12EF28A-AA7C-4965-862A-77A97840C7BD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C12EF28A-AA7C-4965-862A-77A97840C7BD}.Release|x64.ActiveCfg = Release|Any CPU
-		{C12EF28A-AA7C-4965-862A-77A97840C7BD}.Release|x64.Build.0 = Release|Any CPU
-		{C12EF28A-AA7C-4965-862A-77A97840C7BD}.Release|x86.ActiveCfg = Release|Any CPU
-		{C12EF28A-AA7C-4965-862A-77A97840C7BD}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
